ci-refactor: pretty major rewrite of members/userprofile handling
still needs to be cleaned up a little more, but in general we
are getting closer to a cleaner API.

+++ entities/tokens.go
@@ -14,7 +14,7 @@ func init() {
 	database.RegisterBucket(TokenBucketName)
 }
 
-// has returns true if the given token is in the database.
+// hasToken returns true if the given token is in the database.
 // Tokens are hashed before they are stored in the database to protect the on
 // disk version of the tokens.
 func hasToken(token string) bool {
@@ -22,31 +22,43 @@ func hasToken(token string) bool {
 	return database.Has(TokenBucketName, fmt.Sprintf("%x", hash))
 }
 
-// get returns the user name associated with the given token, if the given
+// getToken returns the user name associated with the given token, if the given
 // token exists in the database. An error is returned if the token is not
 // present in the database, or the database operation failed.
 // Tokens are hashed before they are stored in the database to protect the on
 // disk version of the tokens.
 func getToken(token string) (user string, err error) {
+	if token == "" {
+		// empty tokens are not allowed
+		return "", ErrNoAccessToken
+	}
 	hash := sha256.Sum256([]byte(token))
 	err = database.Get(TokenBucketName, fmt.Sprintf("%x", hash), &user)
 	return user, err
 }
 
-// put stores the association between the given token and user name in the
+// putToken stores the association between the given token and user name in the
 // database. An error is returned if the database operation failed.
 // Tokens are hashed before they are stored in the database to protect the on
 // disk version of the tokens.
 func putToken(token, user string) error {
+	if token == "" {
+		// empty tokens are not allowed
+		return ErrNoAccessToken
+	}
 	hash := sha256.Sum256([]byte(token))
 	return database.Put(TokenBucketName, fmt.Sprintf("%x", hash), user)
 }
 
-// remove removes the given token from the database. An error is returned if the
-// database operation failed.
+// removeToken removes the given token from the database. An error is returned
+// if the database operation failed.
 // Tokens are hashed before they are stored in the database to protect the on
 // disk version of the tokens.
 func removeToken(token string) error {
+	if token == "" {
+		// empty tokens are not allowed
+		return ErrNoAccessToken
+	}
 	hash := sha256.Sum256([]byte(token))
 	return database.Remove(TokenBucketName, fmt.Sprintf("%x", hash))
 }
